Move server side props to page - break build
